Add login and password when creating a new user

+++ Website/Controllers/UserController.cs
@@ -1,12 +1,16 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using NuGet.Protocol.Plugins;
+using System.Data;
 using Website.Database;
 using Website.Entities;
+using Website.Models.Account;
+using Website.Service.AccountService;
 
 namespace Website.Controllers
 {
     public class UserController : Controller
     {
-        private  AppDbContext _context;
+        private AppDbContext _context;
 
         public UserController(AppDbContext context)
         {
@@ -19,27 +23,57 @@ namespace Website.Controllers
             return View(users);
         }
 
-        public IActionResult Create() 
+        public IActionResult Create()
         {
             return View();
         }
 
-        public IActionResult Add(User user)
+        [HttpPost]
+        public IActionResult Add(Register register)
         {
-            if (ModelState.IsValid)
+            User user = new User
             {
-                _context.Users.Add(user);
-                _context.SaveChanges();
-                _context.Dispose();
-                return RedirectToAction("Index");
+                Name = register.User.Name,
+                LastName = register.User.LastName,
+                Email = register.User.Email
+            };
+
+            _context.Users.Add(user);
+            _context.SaveChanges();
+
+            Role role = (Role)_context.Roles.Where(i => i.Id == 3);
+
+            if (role == null)
+            {
+                Console.WriteLine("Rola o Id=3 nie istnieje w bazie danych.");
+                return RedirectToAction("Create");
             }
 
-            return RedirectToAction("Create");
+            //if (ModelState.IsValid)
+            //{
+
+            LoginData loginData = new LoginData
+            (
+                register.LoginData.Login,
+                register.LoginData.Password,
+                user.Id,
+                user,
+                role.Id,
+                role
+            );
+
+            _context.LoginData.Add(loginData);
+            _context.SaveChanges();
+            _context.Dispose();
+            return RedirectToAction("Index");
+            //}
+
+            //return View("Create");
         }
 
-        public IActionResult Edit(int Id) 
+        public IActionResult Edit(int Id)
         {
-            User editUser = _context.Users.Where(i => i.Id == Id).FirstOrDefault();
+            User editUser = (User)_context.Users.Where(i => i.Id == Id);
 
             _context.Dispose();
             return View(editUser);
@@ -47,7 +81,7 @@ namespace Website.Controllers
 
         public IActionResult Save(User user)
         {
-            User editUser = _context.Users.Where(i => i.Id == user.Id).FirstOrDefault();
+            User editUser = (User)_context.Users.Where(i => i.Id == user.Id);
 
             editUser.Name = user.Name;
             editUser.LastName = user.LastName;
@@ -59,7 +93,7 @@ namespace Website.Controllers
         }
         public IActionResult Details(int Id)
         {
-            User detailsUser = _context.Users.Where(i => i.Id == Id).FirstOrDefault();
+            User detailsUser = (User)_context.Users.Where(i => i.Id == Id);
 
             _context.Dispose();
             return View(detailsUser);
@@ -67,7 +101,7 @@ namespace Website.Controllers
 
         public IActionResult Delete(int Id)
         {
-            User deleteUser = _context.Users.Where(i => i.Id == Id).FirstOrDefault();
+            User deleteUser = (User)_context.Users.Where(i => i.Id == Id);
 
             _context.Dispose();
             return View(deleteUser);
@@ -75,13 +109,22 @@ namespace Website.Controllers
 
         public IActionResult Remove(User user)
         {
-            User deleteUser = _context.Users.Where(i => i.Id == user.Id).FirstOrDefault();
+            User deleteUser = (User)_context.Users.Where(i => i.Id == user.Id);
+            LoginData loginData = (LoginData)_context.LoginData.Where(i => i.Id_User == user.Id);
 
-            _context.Users.Remove(deleteUser);
-            _context.SaveChanges();
+            if (deleteUser != null && loginData != null)
+            {
+                _context.Users.Remove(deleteUser);
+                _context.LoginData.Remove(loginData);
+                _context.SaveChanges();
+            }
+            else
+            {
+                Console.WriteLine("Nie można usunąć użytkownika. Brak danych w bazie.");
+            }
+
             _context.Dispose();
-            return Redirect("index");
+            return RedirectToAction("Index");
         }
     }
-
 }
